+++ js/form.js
@@ -1,5 +1,5 @@
 import { isEscapeKey } from './util.js';
-import { COMMENT_MAX_LENGTH, HASHTAGS_MAX } from './constants.js';
+import { isValid, resetValidation } from './validation.js';
 
 const uploadForm = document.querySelector('.img-upload__form');
 const fileUpload = document.querySelector('#upload-file');
@@ -8,6 +8,7 @@ const closeButton = document.querySelector('#upload-cancel');
 const tagsField = document.querySelector('.text__hashtags');
 const photoComment = document.querySelector('.text__description');
 
+
 fileUpload.addEventListener('change', () => {
   popupUpload.classList.remove('hidden');
   document.querySelector('body').classList.add('modal-open');
@@ -15,9 +16,8 @@ fileUpload.addEventListener('change', () => {
 });
 
 const resetForm = () => {
-  fileUpload.value = '';
-  tagsField.value = '';
-  photoComment.value = '';
+  uploadForm.reset();
+  resetValidation();
 };
 
 const closeForm = () => {
@@ -31,72 +31,18 @@ closeButton.addEventListener('click', () => {
   closeForm();
 });
 
+uploadForm.addEventListener('submit', (evt) => {
+  if(!isValid()){
+    evt.preventDefault();
+  }
+});
+
 function onDocumentKeydown(evt) {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     if (document.activeElement === photoComment || document.activeElement === tagsField) {
-      return;
+      return document.activeElement.blur();
     }
     closeForm();
   }
 }
-
-const pristine = new Pristine(uploadForm, {
-  classTo: 'img-upload__text',
-  errorClass: 'img-upload__text--invalid',
-  successClass: 'img-upload__text--valid',
-  errorTextParent: 'img-upload__text',
-  errorTextTag: 'span',
-  errorTextClass: 'form__error'
-});
-
-// function validateComment (value) {
-//   return value.length <= COMMENT_MAX_LENGTH;
-// }
-
-pristine.addValidator(
-  photoComment,
-  (value) => {
-    if (value.length <= COMMENT_MAX_LENGTH) {
-      return true;
-    }
-    return 'слишком длинный коммени';
-  });
-
-const hashtagValidations = {
-  startsWithHash: {
-    test: (tag) => tag.startsWith('#'),
-    errorMessage: 'Хэштег должен начинаться с символа #'
-  },
-  validCharacters: {
-    test: (tag) => /^#[a-zа-яё0-9]{1,19}$/i.test(tag),
-    errorMessage: 'Строка после решётки должна состоять из букв и чисел и не может содержать пробелы, спецсимволы и т.д.'
-  },
-  maxLength: {
-    test: (tag) => tag.length <= 20,
-    errorMessage: 'Максимальная длина одного хэш-тега 20 символов, включая решётку'
-  },
-  uniqueTags: {
-    test: (tags) => new Set(tags).size === tags.length,
-    errorMessage: 'Один и тот же хэш-тег не может быть использован дважды'
-  },
-  maxTags: {
-    test: (tags) => tags.length <= HASHTAGS_MAX,
-    errorMessage: 'Нельзя указать больше пяти хэш-тегов'
-  }
-};
-
-Object.keys(hashtagValidations).forEach((key) => {
-  pristine.addValidator(
-    tagsField,
-    (value) => {
-      const hashtags = value.trim().split(/\s+/).filter((tag) => tag !== '');
-      if (key === 'uniqueTags' || key === 'maxTags') {
-        return hashtagValidations[key].test(hashtags);
-      }
-      return hashtags.every((tag) => hashtagValidations[key].test(tag));
-    },
-    hashtagValidations[key].errorMessage
-  );
-});
-
